+++ Libs/rdroid/Primitives/rdParticle.h
@@ -10,8 +10,8 @@ typedef rdParticle* (J3DAPI* rdParticleLoaderFunc)(const char* pFilename);
 
 rdParticleLoaderFunc J3DAPI rdParticle_RegisterLoader(rdParticleLoaderFunc pFunc); // Added
 
-rdParticle* J3DAPI rdParticle_New(unsigned int num, float size, rdMaterial* pMat, rdLightMode lightMode);
-int J3DAPI rdParticle_NewEntry(rdParticle* pParticle, unsigned int num, float size, rdMaterial* pMaterial, rdLightMode lightMode);
+rdParticle* J3DAPI rdParticle_New(size_t num, float size, rdMaterial* pMat, rdLightMode lightMode);
+int J3DAPI rdParticle_NewEntry(rdParticle* pParticle, size_t num, float size, rdMaterial* pMaterial, rdLightMode lightMode);
 rdParticle* J3DAPI rdParticle_Duplicate(const rdParticle* pOriginal);
 
 void J3DAPI rdParticle_Free(rdParticle* pParticle);
